# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TN Bus+ API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/depots:
    get:
      operationId: listDepots
      tags: [admin]
      summary: List all depots
      responses:
        "200":
          description: Depots list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Depot"
    post:
      operationId: createDepot
      tags: [admin]
      summary: Create a depot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepotInput"
      responses:
        "201":
          description: Depot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Depot"
        "400":
          description: Invalid input
  /admin/depots/{id}:
    patch:
      operationId: updateDepot
      tags: [admin]
      summary: Update a depot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepotInput"
      responses:
        "200":
          description: Depot updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Depot"
        "400":
          description: Invalid input
        "404":
          description: Not found
    delete:
      operationId: deleteDepot
      tags: [admin]
      summary: Delete a depot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Depot deleted
        "404":
          description: Not found
        "409":
          description: Depot in use
  /admin/crew:
    get:
      operationId: listCrew
      tags: [admin]
      summary: List all crew members
      responses:
        "200":
          description: Crew list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Crew"
    post:
      operationId: createCrew
      tags: [admin]
      summary: Create a crew member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrewInput"
      responses:
        "201":
          description: Crew created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crew"
        "400":
          description: Invalid input
  /admin/crew/{id}:
    patch:
      operationId: updateCrew
      tags: [admin]
      summary: Update a crew member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrewInput"
      responses:
        "200":
          description: Crew updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Crew"
        "400":
          description: Invalid input
        "404":
          description: Not found
    delete:
      operationId: deleteCrew
      tags: [admin]
      summary: Delete a crew member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Crew deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    Depot:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        city:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        manager:
          type: ["string", "null"]
        capacity:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - code
        - city
        - latitude
        - longitude
        - capacity
        - createdAt
    DepotInput:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        city:
          type: string
        manager:
          type: ["string", "null"]
        capacity:
          type: integer
      required:
        - name
        - code
        - city
    Crew:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        depotId:
          type: ["integer", "null"]
        phone:
          type: ["string", "null"]
        licenseNumber:
          type: ["string", "null"]
        status:
          type: string
        experienceYears:
          type: integer
        safetyScore:
          type: number
        assignedBusNumber:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - role
        - status
        - experienceYears
        - safetyScore
        - createdAt
    CrewInput:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        depotId:
          type: ["integer", "null"]
        phone:
          type: ["string", "null"]
        licenseNumber:
          type: ["string", "null"]
        status:
          type: string
        experienceYears:
          type: integer
        safetyScore:
          type: number
        assignedBusNumber:
          type: ["string", "null"]
      required:
        - name
        - role
...