# lib/api-zod/src/generated/types/crew.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TN Bus+ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Crew {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  depotId?: number | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  licenseNumber?: string | null;
  status: string;
  experienceYears: number;
  safetyScore: number;
  /** @nullable */
  assignedBusNumber?: string | null;
  createdAt: string;
}