# lib/api-zod/src/generated/types/depot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TN Bus+ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Depot {
  id: number;
  name: string;
  code: string;
  city: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  manager?: string | null;
  capacity: number;
  createdAt: string;
}